add enums definitions file to lib directory
