+++ src/index.ts
@@ -3,3 +3,4 @@ export { default as directUpload } from './lib/directUpload';
 export { default as ActiveStorageProvider } from './ActiveStorageProvider';
 export { default as DirectUpload } from './DirectUpload';
 export { default as useDirectUpload } from './useDirectUpload';
+export { DirectUploadResultStatus } from './lib/enums';
