+++ src/lib/enums.ts
@@ -1,0 +1,7 @@
+export enum DirectUploadResultStatus {
+  success = 'success',
+  uploading = 'uploading',
+  error = 'error',
+  waiting = 'waiting',
+  canceled = 'canceled'
+}
